hello_world/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.2
colorama==0.4.4
distlib==0.3.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
isort==5.8.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
pylint==2.7.4
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.3
wrapt==1.12.1
